+++ ComplexNumber.h
@@ -6,10 +6,10 @@ class ComplexNumber {
     static const double ARBRITARY_SMALL_NUM;
     static ComplexNumber dZ, z;
 public:
-    double a, b;
-    ComplexNumber(double a, double b);
-    int getNumEscapeSteps(const ComplexNumber& c);
-    double magnitudeSquared();
+    long double a, b;
+    ComplexNumber(long double a, long double b);
+    short getNumEscapeSteps(const ComplexNumber& c);
+    long double magnitudeSquared();
     ComplexNumber operator+(const ComplexNumber& rhs) const;
     ComplexNumber operator*(const ComplexNumber& rhs) const;
     ComplexNumber operator*(const int lhs) const;
